Factor out common class for editor screen settings
